nlp-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-dateutil==2.8.2
regex==2023.12.25
pytest==7.4.0
httpx==0.26.0
